+++ src/u-config.c
@@ -600,7 +600,7 @@ static s8 *fieldbyid(pkg *p, i32 id)
 {
     assert(id >= 0);
     assert(id < PKG_NFIELDS);
-    return (s8 *)((byte *)&p->name + id*sizeof(s8));
+    return (s8 *)((byte *)&p->name + id*(i32)sizeof(s8));
 }
 
 static s8 *fieldbyname(pkg *p, s8 name)
